메시지 삭제 기능 구현 및 패키지 정리
