+++ src/main/java/com/project/wah/projectwebservice/domain/message/Message.java
@@ -6,13 +6,8 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.hibernate.annotations.OnDelete;
-import org.hibernate.annotations.OnDeleteAction;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.format.annotation.DateTimeFormat;
 
 import javax.persistence.*;
-import java.time.LocalDate;
 import java.time.LocalDateTime;
 
 @Getter
@@ -25,36 +20,36 @@ public class Message {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "MESSAGE_ID", nullable = false)
-    private int id;
+    private Long id;
 
-    // 쪽지 송신하는 사용자
+    // 메시지 송신하는 사용자
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "sender_id")
+    @JoinColumn(name = "sender_id", nullable = false)
     @JsonManagedReference   // 순환참조 방지
     // @OnDelete(action = OnDeleteAction.NO_ACTION)    // 작성자 혹은 수신자가 계정을 삭제하면 같이 지워짐
     private User sender;
 
-    // 쪽지 수신하는 사용자
+    // 메시지 수신하는 사용자
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "receiver_id")
+    @JoinColumn(name = "receiver_id", nullable = false)
     @JsonManagedReference   // 순환참조 방지
     // @OnDelete(action = OnDeleteAction.NO_ACTION)    // 작성자 혹은 수신자가 계정을 삭제하면 같이 지워짐
     private User receiver;
 
+    // 메시지 제목
     @Column(nullable = false)
     private String title;
 
-    @Column(nullable = false)
+    // 메시지 내용, 메시지 내용은 최대 500자로 설정
+    @Column(columnDefinition = "TEXT",length = 500, nullable = false)
     private String content;
 
-    @Column(nullable = false)
-    private boolean readstate;
-
+    // 메시지를 생성한 날짜
     @Column(nullable = false)
     private LocalDateTime createdate;
 
     public void sendUser(User user) { this.sender = user; }
-    public void changeReadState() { this.readstate = true; }
+
