+++ src/main/java/com/project/wah/projectwebservice/service/MessageService.java
@@ -23,7 +23,7 @@ public class MessageService {
     private final MessageRepository messageRepository;
     private final UserRepository userRepository;
 
-    // 쪽지 보내기
+    // 메시지 보내기
     @Transactional
     public Message createMessage(MessageCreateRequestDto requestDto, SessionUser sessionUser) {
 
@@ -34,7 +34,7 @@ public class MessageService {
 
         User receivedUser = receivedUserOptional.get();
 
-        // 메시지 보내는 코드
+        // 메시지 생성
         Message message = Message.builder()
                 .receiver(receivedUser)
                 .title(requestDto.getTitle())
@@ -47,39 +47,16 @@ public class MessageService {
 
     }
 
-    // 쪽지 상세 읽기
-    public MessageReadResponseDto findById(int id) {
+    // 메시지 상세 읽기
+    @Transactional
+    public MessageReadResponseDto findById(Long id, SessionUser sessionUser) {
         Message message = messageRepository.findById(id)
                 .orElseThrow(() -> new IllegalArgumentException("해당 메시지가 없습니다. id= " + id));
 
         return new MessageReadResponseDto(message);
     }
 
-
-//    // 쪽지 삭제
-//    @Transactional
-//    public ResponseEntity<MessageDeleteResponseDto> deleteMessage(int messageId, @LoginUser SessionUser sessionUser) {
-//
-//        Optional<User> userOptional = userRepository.findById(sessionUser.getId());
-//        User user = userOptional.get();
-//
-//        boolean isValid = user.deleteMessage(messageId);
-//
-//        if (!isValid) {
-//            log.error("nickname={}, messageId={}, error={}", user.getNickname(), messageId, "해당 쪽지를 찾을 수 없음");
-//            return new ResponseEntity<>(
-//                    MessageDeleteResponseDto.builder().status(StatusMessage.BAD_REQUEST).build(),
-//                    HttpStatus.valueOf(StatusCode.BAD_REQUEST)
-//            );
-//        }
-//        messageRepository.deleteById(messageId);
-//
-//        return new ResponseEntity<>(
-//                MessageDeleteResponseDto.builder().status(StatusMessage.SUCCESS).build(),
-//                HttpStatus.valueOf(StatusCode.SUCCESS)
-//        );
-//    }
-    // 받은 쪽지 리스트 조회
+    // 받은 메시지 리스트 조회
     @Transactional
     public Page<MessageListReadResponseDto> findAllReceiverDesc(@LoginUser SessionUser sessionUser, Pageable pageable) {
 
@@ -90,7 +67,8 @@ public class MessageService {
         return messagePagingList;
     }
 
-    // 보낸 쪽지 리스트 조회
+    // 보낸 메시지 리스트 조회
+    @Transactional
     public Page<MessageListReadResponseDto> findAllSenderDesc(@LoginUser SessionUser sessionUser, Pageable pageable) {
 
         Page<Message> messageList = messageRepository.findAllSenderDesc(sessionUser.getId(), pageable);
@@ -100,68 +78,11 @@ public class MessageService {
         return messagePagingList;
     }
 
-//    // 받은 쪽지 검색 후 리스트 조회
-//    @Transactional
-//    public Page<MessageListReadResponseDto> findReceiverMessageSearch(String title, Pageable pageable, @LoginUser SessionUser sessionUser) {
-//
-//        Page<Message> messageList = messageRepository.findReceiverByTitleContaining(title, pageable);
-//
-//        Page<MessageListReadResponseDto> messagePagingList = messageList.map(message -> new MessageListReadResponseDto(message));
-//
-//        return messagePagingList;
-//    }
-//
-//    // 보낸 쪽지 검색 후 리스트 조회
-//    @Transactional
-//    public Page<MessageListReadResponseDto> findSenderMessagesearch(@LoginUser SessionUser sessionUser, Pageable pageable) {
-//
-//        Page<Message> messageList = messageRepository.findSenderByTitleContaining(sessionUser.getId(), pageable);
-//
-//        Page<MessageListReadResponseDto> messagePagingList = messageList.map(message -> new MessageListReadResponseDto(message));
-//
-//        return messagePagingList;
-//    }
+    // 메시지 삭제
+    public void deleteMessage (Long id) {
+        Message message = messageRepository.findById(id)
+                .orElseThrow(() -> new IllegalArgumentException("해당 메시지가 없습니다. id= " + id));
 
-
-
+        messageRepository.delete(message);
+    }
 }
-
-
-//    Optional<User> userOptional = userRepository.findById(sessionUser.getId());
-//    User sendUser = userOptional.get();
-//
-//    Optional<User> receivedUserOptional = userRepository.findByNickname(requestDto.getReceiver());
-//        if(receivedUserOptional.isEmpty()) {
-//                log.error("nickname={}, error={}", requestDto.getReceiver(), "쪽지 수신자가 존재하지 않음");
-//                return new ResponseEntity<>(
-//        MessageCreateResponseDto.builder().status(StatusMessage.UNKNOWN_RECEIVER).build(),
-//        HttpStatus.valueOf(StatusCode.BAD_REQUEST)
-//        );
-//        };
-//
-//        User receivedUser = receivedUserOptional.get();
-//
-//        // 쪽지 내용 255자 이내로 제한
-//        if (requestDto.getContent().length() > 255) {
-//        log.error("nickname={}, error={}", sendUser.getNickname(), "쪽지 내용 255자 초과");
-//        return new ResponseEntity<>(
-//        MessageCreateResponseDto.builder().status(StatusMessage.BAD_REQUEST).build(),
-//        HttpStatus.valueOf(StatusCode.BAD_REQUEST)
-//        );
-//        }
-//
-//        // 메시지 보내는 코드
-//        Message message = Message.builder()
-//        .receiver(receivedUser)
-//        .title(requestDto.getTitle())
-//        .content(requestDto.getContent())
-//        .createdate(requestDto.getCreateDate())
-//        .build();
-//
-//        sendUser.sendMessage(message);
-//        messageRepository.save(message);
-//
-//        return new ResponseEntity<>(
-//        MessageCreateResponseDto.builder().status(StatusMessage.SUCCESS).build(),
-//        HttpStatus.valueOf(StatusCode.SUCCESS)
-//        );
