--- src/main/java/com/project/wah/projectwebservice/web/dto/message/MessageDeleteResponseDto.java
@@ -1,10 +1,0 @@
-package com.project.wah.projectwebservice.web.dto.message;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Builder
-@Getter
-public class MessageDeleteResponseDto {
-    private String status;
-}
