+++ src/main/resources/static/js/app/user.js
@@ -1,9 +1,6 @@
 var main = {
     init : function () {
         var _this = this;
-        $('#btn-save').on('click', function () {
-            _this.save();
-        });
 
         $('#btn-update').on('click', function() {
             _this.update();
@@ -13,29 +10,6 @@ var main = {
             _this.delete();
         });
 
-    },
-    save : function () {
-        var data = {
-            nickname: $('#nickname').val(),
-            aboutme: $('#aboutme').val(),
-            mentoremail: $('#mentoremail').val(),
-            githubemail: $('#githubemail').val(),
-            blogaddress: $('#blogaddress').val(),
-            phonenumber: $('#phonenumber').val()
-        };
-
-        $.ajax({
-            type: 'POST',
-            url: '/api/v1/userdetails',
-            dataType: 'json',
-            contentType:'application/json; charset=utf-8',
-            data: JSON.stringify(data)
-        }).done(function() {
-            alert('글이 등록되었습니다.');
-            window.location.href = '/';
-        }).fail(function (error) {
-            alert(JSON.stringify(error));
-        });
     },
     update : function() {
         var data = {
