+++ src/main/resources/templates/layout/footer.mustache
@@ -3,7 +3,7 @@
 <script src="https://code.jquery.com/jquery-3.3.1.min.js"></script>
 <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/js/bootstrap.min.js"></script>
 
-<script src="/js/app/index.js"></script>
+<script src="/js/app/user.js"></script>
 <script src="/js/app/message.js"></script>
 </body>
 </html>
